fix: Change type of currency factor to float.

+++ order.go
@@ -55,7 +55,7 @@ type OrdersReturn struct {
 	Data  []struct {
 		OrderNumber      string    `json:"orderNumber"`
 		CurrencyId       string    `json:"currencyId"`
-		CurrencyFactor   int       `json:"currencyFactor"`
+		CurrencyFactor   float64   `json:"currencyFactor"`
 		SalesChannelId   string    `json:"salesChannelId"`
 		BillingAddressId string    `json:"billingAddressId"`
 		OrderDateTime    time.Time `json:"orderDateTime"`
@@ -256,7 +256,7 @@ type OrderReturn struct {
 	Data struct {
 		OrderNumber      string    `json:"orderNumber"`
 		CurrencyId       string    `json:"currencyId"`
-		CurrencyFactor   int       `json:"currencyFactor"`
+		CurrencyFactor   float64   `json:"currencyFactor"`
 		SalesChannelId   string    `json:"salesChannelId"`
 		BillingAddressId string    `json:"billingAddressId"`
 		OrderDateTime    time.Time `json:"orderDateTime"`
